+++ packages/core/app-layout/src/components/pageHeader/AppPageHeader.vue
@@ -1,7 +1,10 @@
 <template>
-  <div class="kong-ui-app-page-header">
+  <div
+    class="kong-ui-app-page-header"
+    :class="{ 'konnect-navigation-next': konnectNavigationNext }"
+  >
     <div
-      v-if="hasBreadcrumbs"
+      v-if="hasBreadcrumbs && !konnectNavigationNext"
       class="page-header-breadcrumbs"
       data-testid="page-header-breadcrumbs"
     >
@@ -21,21 +24,22 @@
     <div class="page-header-title-section">
       <div class="page-header-title-wrapper">
         <div
-          v-if="$slots['title-before']"
+          v-if="$slots['title-before'] && !konnectNavigationNext"
           class="page-header-title-before"
           data-testid="page-header-title-before"
         >
           <slot name="title-before" />
         </div>
-        <h1
+        <component
+          :is="konnectNavigationNext ? 'h2' : 'h1'"
           class="page-header-title"
           data-testid="page-header-title"
           :title="title"
         >
           {{ title }}
-        </h1>
+        </component>
         <div
-          v-if="$slots['title-after']"
+          v-if="$slots['title-after'] && !konnectNavigationNext"
           class="page-header-title-after"
           data-testid="page-header-title-after"
         >
@@ -75,6 +79,11 @@ const props = defineProps({
   breadcrumbs: {
     type: Array as PropType<BreadcrumbItem[]>,
     default: () => ([]),
+  },
+  /** Temporary prop for Konnect navigation next. This will be removed when the Konnect navigation next is fully implemented. */
+  konnectNavigationNext: {
+    type: Boolean,
+    default: false,
   },
 })
 
@@ -145,6 +154,22 @@ const breadcrumbIconSlots = computed((): string[] => {
 
     .page-header-section-below {
       margin-top: unset;
+    }
+  }
+
+  &.konnect-navigation-next {
+    .page-header-title-section {
+      .page-header-title {
+        font-size: var(--kui-font-size-50, $kui-font-size-50);
+        line-height: var(--kui-line-height-40, $kui-line-height-40);
+      }
+    }
+
+    .page-header-section-below {
+      color: var(--kui-color-text-neutral, $kui-color-text-neutral);
+      font-size: var(--kui-font-size-30, $kui-font-size-30);
+      font-weight: var(--kui-font-weight-regular, $kui-font-weight-regular);
+      line-height: var(--kui-line-height-30, $kui-line-height-30);
     }
   }
 }