+++ packages/entities/entities-plugins/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.159.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.159.1...@kong-ui-public/entities-plugins@9.159.2) (2026-02-27)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+## [9.159.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.159.0...@kong-ui-public/entities-plugins@9.159.1) (2026-02-27)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
+# [9.159.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.158.5...@kong-ui-public/entities-plugins@9.159.0) (2026-02-26)
+
+
+### Features
+
+* **entities-plugins:** add metadata and icons for 3.14 new plugins ([#2943](https://github.com/Kong/public-ui-components/issues/2943)) ([f81ff14](https://github.com/Kong/public-ui-components/commit/f81ff14df038ac054b750d0fc5ba8e8744edc53d))
+
+
+
+
+
+## [9.158.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.158.4...@kong-ui-public/entities-plugins@9.158.5) (2026-02-26)
+
+**Note:** Version bump only for package @kong-ui-public/entities-plugins
+
+
+
+
+
 ## [9.158.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-plugins@9.158.3...@kong-ui-public/entities-plugins@9.158.4) (2026-02-25)
 
 **Note:** Version bump only for package @kong-ui-public/entities-plugins