+++ packages/entities/entities-redis-configurations/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@1.9.19...@kong-ui-public/entities-redis-configurations@1.9.20) (2026-02-27)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
+## [1.9.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@1.9.18...@kong-ui-public/entities-redis-configurations@1.9.19) (2026-02-26)
+
+**Note:** Version bump only for package @kong-ui-public/entities-redis-configurations
+
+
+
+
+
 ## [1.9.18](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-redis-configurations@1.9.17...@kong-ui-public/entities-redis-configurations@1.9.18) (2026-02-12)
 
 **Note:** Version bump only for package @kong-ui-public/entities-redis-configurations